RAG-upload/requirements.txt
python-dotenv>=1.0.0
pinecone-client>=3.0.0
pinecone-plugin-assistant>=0.1.0
numpy>=1.20.0
pandas>=1.3.0
requests>=2.28.0
